+++ Gruntfile.js
@@ -71,6 +71,7 @@ var JSUnits = {
 			'src/js/molpad/MPBond.js',
 			'src/js/molpad/MPBond_calc.js',
 			'src/js/molpad/MPBond_handler.js',
+			'src/js/molpad/MolPad_settings.js',
 			'src/js/molpad/MolPad.js',
 			'src/js/molpad/MPMolecule.js',
 			'src/js/molpad/MPSelection.js',